Documentation for Set Hint
It seems clear that we need a concept other than `group` to
describe things like this:

  - a set of options
  - a set of input fields
  - a set of controls
  - a set of product attributes

I've been experimenting with this concept for a while, and
it's been extremely helpful in simplifying the task of
creating meaningful and useful apps with a small amount of
information.

We've discussed the term `collection` to convey something like
this, but I think `set` more closely matches the natural way
we've named these things in the past, it's shorter, and I think
it gets to something about an emphasis on similarity that
`collection` misses.